Outsource context into separate provider
